+++ packages/remix-dev/config.ts
@@ -4,16 +4,16 @@ import { pathToFileURL } from "node:url";
 import fse from "fs-extra";
 import getPort from "get-port";
 import NPMCliPackageJson from "@npmcli/package-json";
+import { coerce } from "semver";
 
 import type { RouteManifest, DefineRoutesFunction } from "./config/routes";
 import { defineRoutes } from "./config/routes";
 import { defineConventionalRoutes } from "./config/routesConvention";
 import { ServerMode, isValidServerMode } from "./config/serverModes";
-import { serverBuildVirtualModule } from "./compiler/virtualModules";
-import { writeConfigDefaults } from "./compiler/utils/tsconfig/write-config-defaults";
+import { serverBuildVirtualModule } from "./compiler/server/virtualModules";
 import { flatRoutes } from "./config/flat-routes";
-import { getPreferredPackageManager } from "./cli/getPreferredPackageManager";
-import { warnOnce } from "./compiler/warnings";
+import { detectPackageManager } from "./cli/detectPackageManager";
+import { warnOnce } from "./warnOnce";
 
 export interface RemixMdxConfig {
   rehypePlugins?: any[];
@@ -37,25 +37,26 @@ export type ServerModuleFormat = "esm" | "cjs";
 export type ServerPlatform = "node" | "neutral";
 
 type Dev = {
-  port?: number;
-  appServerPort?: number;
-  remixRequestHandlerPath?: string;
-  rebuildPollIntervalMs?: number;
-};
+  port?: number; // TODO: remove in v2
 
-export type VanillaExtractOptions = {
-  cache?: boolean;
+  command?: string;
+  httpScheme?: string;
+  httpHost?: string;
+  httpPort?: number;
+  websocketPort?: number;
+  restart?: boolean;
+  publicDirectory?: string;
 };
 
 interface FutureConfig {
-  unstable_cssModules: boolean;
-  unstable_cssSideEffectImports: boolean;
   unstable_dev: boolean | Dev;
+  /** @deprecated Use the `postcss` config option instead */
   unstable_postcss: boolean;
+  /** @deprecated Use the `tailwind` config option instead */
   unstable_tailwind: boolean;
-  unstable_vanillaExtract: boolean | VanillaExtractOptions;
   v2_errorBoundary: boolean;
   v2_meta: boolean;
+  v2_normalizeFormMethod: boolean;
   v2_routeConvention: boolean;
 }
 
@@ -119,6 +120,12 @@ export interface AppConfig {
    * Additional MDX remark / rehype plugins.
    */
   mdx?: RemixMdxConfig | RemixMdxConfigFunction;
+
+  /**
+   * Whether to process CSS using PostCSS if `postcss.config.js` is present.
+   * Defaults to `false`.
+   */
+  postcss?: boolean;
 
   /**
    * A server entrypoint, relative to the root directory that becomes your
@@ -187,6 +194,12 @@ export interface AppConfig {
    * The platform the server build is targeting. Defaults to "node".
    */
   serverPlatform?: ServerPlatform;
+
+  /**
+   * Whether to support Tailwind functions and directives in CSS files if `tailwindcss` is installed.
+   * Defaults to `false`.
+   */
+  tailwind?: boolean;
 
   /**
    * A list of filenames or a glob patterns to match files in the `app/routes`
@@ -281,6 +294,12 @@ export interface RemixConfig {
   mdx?: RemixMdxConfig | RemixMdxConfigFunction;
 
   /**
+   * Whether to process CSS using PostCSS if `postcss.config.js` is present.
+   * Defaults to `false`.
+   */
+  postcss: boolean;
+
+  /**
    * The path to the server build file. This file should end in a `.js`.
    */
   serverBuildPath: string;
@@ -348,6 +367,12 @@ export interface RemixConfig {
   serverPlatform: ServerPlatform;
 
   /**
+   * Whether to support Tailwind functions and directives in CSS files if `tailwindcss` is installed.
+   * Defaults to `false`.
+   */
+  tailwind: boolean;
+
+  /**
    * A list of directories to watch.
    */
   watchPaths: string[];
@@ -401,12 +426,20 @@ export async function readConfig(
     }
   }
 
+  if (appConfig.serverBuildTarget) {
+    warnOnce(serverBuildTargetWarning, "v2_serverBuildTarget");
+  }
+
   if (!appConfig.future?.v2_errorBoundary) {
     warnOnce(errorBoundaryWarning, "v2_errorBoundary");
   }
 
-  if (appConfig.serverBuildTarget) {
-    warnOnce(serverBuildTargetWarning, "v2_serverBuildTarget");
+  if (!appConfig.future?.v2_normalizeFormMethod) {
+    warnOnce(formMethodWarning, "v2_normalizeFormMethod");
+  }
+
+  if (!appConfig.future?.v2_meta) {
+    warnOnce(metaWarning, "v2_meta");
   }
 
   let isCloudflareRuntime = ["cloudflare-pages", "cloudflare-workers"].includes(
@@ -424,6 +457,11 @@ export async function readConfig(
   let serverEntryPoint = appConfig.server;
   let serverMainFields = appConfig.serverMainFields;
   let serverMinify = appConfig.serverMinify;
+
+  if (!appConfig.serverModuleFormat) {
+    warnOnce(serverModuleFormatWarning, "serverModuleFormatWarning");
+  }
+
   let serverModuleFormat = appConfig.serverModuleFormat || "cjs";
   let serverPlatform = appConfig.serverPlatform || "node";
   if (isCloudflareRuntime) {
@@ -445,7 +483,43 @@ export async function readConfig(
     serverModuleFormat === "esm" ? ["module", "main"] : ["main", "module"];
   serverMinify ??= false;
 
+  if (appConfig.future) {
+    if ("unstable_cssModules" in appConfig.future) {
+      warnOnce(
+        'The "future.unstable_cssModules" config option has been removed as this feature is now enabled automatically.'
+      );
+    }
+
+    if ("unstable_cssSideEffectImports" in appConfig.future) {
+      warnOnce(
+        'The "future.unstable_cssSideEffectImports" config option has been removed as this feature is now enabled automatically.'
+      );
+    }
+
+    if ("unstable_vanillaExtract" in appConfig.future) {
+      warnOnce(
+        'The "future.unstable_vanillaExtract" config option has been removed as this feature is now enabled automatically.'
+      );
+    }
+
+    if (appConfig.future.unstable_postcss !== undefined) {
+      warnOnce(
+        'The "future.unstable_postcss" config option has been deprecated as this feature is now considered stable. Use the "postcss" config option instead.'
+      );
+    }
+
+    if (appConfig.future.unstable_tailwind !== undefined) {
+      warnOnce(
+        'The "future.unstable_tailwind" config option has been deprecated as this feature is now considered stable. Use the "tailwind" config option instead.'
+      );
+    }
+  }
+
   let mdx = appConfig.mdx;
+  let postcss =
+    appConfig.postcss ?? appConfig.future?.unstable_postcss === true;
+  let tailwind =
+    appConfig.tailwind ?? appConfig.future?.unstable_tailwind === true;
 
   let appDirectory = path.resolve(
     rootDirectory,
@@ -471,28 +545,6 @@ export async function readConfig(
   if (userEntryServerFile) {
     entryServerFile = userEntryServerFile;
   } else {
-    if (!deps["isbot"]) {
-      console.log(
-        "adding `isbot` to your package.json, you should commit this change"
-      );
-
-      pkgJson.update({
-        dependencies: {
-          ...pkgJson.content.dependencies,
-          isbot: "latest",
-        },
-      });
-
-      await pkgJson.save();
-
-      let packageManager = getPreferredPackageManager();
-
-      execSync(`${packageManager} install`, {
-        cwd: remixRoot,
-        stdio: "inherit",
-      });
-    }
-
     let serverRuntime = deps["@remix-run/deno"]
       ? "deno"
       : deps["@remix-run/cloudflare"]
@@ -507,27 +559,85 @@ export async function readConfig(
         "@remix-run/cloudflare",
         "@remix-run/node",
       ];
-      let formattedList = listFormat.format(serverRuntimes);
+      let formattedList = disjunctionListFormat.format(serverRuntimes);
       throw new Error(
         `Could not determine server runtime. Please install one of the following: ${formattedList}`
       );
     }
 
-    entryServerFile = `entry.server.${serverRuntime}.tsx`;
+    let clientRenderer = deps["@remix-run/react"] ? "react" : undefined;
+
+    if (!clientRenderer) {
+      throw new Error(
+        `Could not determine renderer. Please install the following: @remix-run/react`
+      );
+    }
+
+    let maybeReactVersion = coerce(deps.react);
+    if (!maybeReactVersion) {
+      let react = ["react", "react-dom"];
+      let list = conjunctionListFormat.format(react);
+      throw new Error(
+        `Could not determine React version. Please install the following packages: ${list}`
+      );
+    }
+
+    let type: "stream" | "string" =
+      maybeReactVersion.major >= 18 || maybeReactVersion.raw === "0.0.0"
+        ? "stream"
+        : "string";
+
+    if (!deps["isbot"] && type === "stream") {
+      console.log(
+        "adding `isbot` to your package.json, you should commit this change"
+      );
+
+      pkgJson.update({
+        dependencies: {
+          ...pkgJson.content.dependencies,
+          isbot: "latest",
+        },
+      });
+
+      await pkgJson.save();
+
+      let packageManager = detectPackageManager() ?? "npm";
+
+      execSync(`${packageManager} install`, {
+        cwd: remixRoot,
+        stdio: "inherit",
+      });
+    }
+
+    entryServerFile = `${serverRuntime}/entry.server.${clientRenderer}-${type}.tsx`;
   }
 
   if (userEntryClientFile) {
     entryClientFile = userEntryClientFile;
   } else {
-    let clientRuntime = deps["@remix-run/react"] ? "react" : undefined;
+    let clientRenderer = deps["@remix-run/react"] ? "react" : undefined;
 
-    if (!clientRuntime) {
+    if (!clientRenderer) {
       throw new Error(
         `Could not determine runtime. Please install the following: @remix-run/react`
       );
     }
 
-    entryClientFile = `entry.client.${clientRuntime}.tsx`;
+    let maybeReactVersion = coerce(deps.react);
+    if (!maybeReactVersion) {
+      let react = ["react", "react-dom"];
+      let list = conjunctionListFormat.format(react);
+      throw new Error(
+        `Could not determine React version. Please install the following packages: ${list}`
+      );
+    }
+
+    let type: "stream" | "string" =
+      maybeReactVersion.major >= 18 || maybeReactVersion.raw === "0.0.0"
+        ? "stream"
+        : "string";
+
+    entryClientFile = `entry.client.${clientRenderer}-${type}.tsx`;
   }
 
   let entryClientFilePath = userEntryClientFile
@@ -537,6 +647,10 @@ export async function readConfig(
   let entryServerFilePath = userEntryServerFile
     ? path.resolve(appDirectory, userEntryServerFile)
     : path.resolve(defaultsDirectory, entryServerFile);
+
+  if (appConfig.browserBuildDirectory) {
+    warnOnce(browserBuildDirectoryWarning, "browserBuildDirectory");
+  }
 
   let assetsBuildDirectory =
     appConfig.assetsBuildDirectory ||
@@ -619,20 +733,13 @@ export async function readConfig(
     tsconfigPath = rootJsConfig;
   }
 
-  if (tsconfigPath) {
-    writeConfigDefaults(tsconfigPath);
-  }
-
   let future: FutureConfig = {
-    unstable_cssModules: appConfig.future?.unstable_cssModules === true,
-    unstable_cssSideEffectImports:
-      appConfig.future?.unstable_cssSideEffectImports === true,
     unstable_dev: appConfig.future?.unstable_dev ?? false,
     unstable_postcss: appConfig.future?.unstable_postcss === true,
     unstable_tailwind: appConfig.future?.unstable_tailwind === true,
-    unstable_vanillaExtract: appConfig.future?.unstable_vanillaExtract ?? false,
     v2_errorBoundary: appConfig.future?.v2_errorBoundary === true,
     v2_meta: appConfig.future?.v2_meta === true,
+    v2_normalizeFormMethod: appConfig.future?.v2_normalizeFormMethod === true,
     v2_routeConvention: appConfig.future?.v2_routeConvention === true,
   };
 
@@ -662,6 +769,8 @@ export async function readConfig(
     serverModuleFormat,
     serverPlatform,
     mdx,
+    postcss,
+    tailwind,
     watchPaths,
     tsconfigPath,
     future,
@@ -722,6 +831,8 @@ const resolveServerBuildPath = (
 
   // retain deprecated behavior for now
   if (appConfig.serverBuildDirectory) {
+    warnOnce(serverBuildDirectoryWarning, "serverBuildDirectory");
+
     serverBuildPath = path.join(appConfig.serverBuildDirectory, "index.js");
   }
 
@@ -732,21 +843,88 @@ const resolveServerBuildPath = (
   return path.resolve(rootDirectory, serverBuildPath);
 };
 
-// @ts-expect-error available in node 12+
-// https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Intl/ListFormat#browser_compatibility
-let listFormat = new Intl.ListFormat("en", {
+// adds types for `Intl.ListFormat` to the global namespace
+// we could also update our `tsconfig.json` to include `lib: ["es2021"]`
+declare namespace Intl {
+  type ListType = "conjunction" | "disjunction";
+
+  interface ListFormatOptions {
+    localeMatcher?: "lookup" | "best fit";
+    type?: ListType;
+    style?: "long" | "short" | "narrow";
+  }
+
+  interface ListFormatPart {
+    type: "element" | "literal";
+    value: string;
+  }
+
+  class ListFormat {
+    constructor(locales?: string | string[], options?: ListFormatOptions);
+    format(values: any[]): string;
+    formatToParts(values: any[]): ListFormatPart[];
+    supportedLocalesOf(
+      locales: string | string[],
+      options?: ListFormatOptions
+    ): string[];
+  }
+}
+
+let conjunctionListFormat = new Intl.ListFormat("en", {
   style: "long",
   type: "conjunction",
 });
 
-export let serverBuildTargetWarning = `⚠️ DEPRECATED: The "serverBuildTarget" config option is deprecated. Use a combination of "publicPath", "serverBuildPath", "serverConditions", "serverDependenciesToBundle", "serverMainFields", "serverMinify", "serverModuleFormat" and/or "serverPlatform" instead.`;
+let disjunctionListFormat = new Intl.ListFormat("en", {
+  style: "long",
+  type: "disjunction",
+});
 
-export let flatRoutesWarning = `⚠️ DEPRECATED: The old nested folders route convention has been deprecated in favor of "flat routes".  Please enable the new routing convention via the \`future.v2_routeConvention\` flag in your \`remix.config.js\` file.  For more information, please see https://remix.run/docs/en/main/file-conventions/route-files-v2.`;
+export let browserBuildDirectoryWarning =
+  "⚠️ REMIX FUTURE CHANGE: The `browserBuildDirectory` config option will be removed in v2. " +
+  "Use `assetsBuildDirectory` instead. " +
+  "For instructions on making this change see " +
+  "https://remix.run/docs/en/v1.15.0/pages/v2#browserbuilddirectory";
+
+export let serverBuildDirectoryWarning =
+  "⚠️ REMIX FUTURE CHANGE: The `serverBuildDirectory` config option will be removed in v2. " +
+  "Use `serverBuildPath` instead. " +
+  "For instructions on making this change see " +
+  "https://remix.run/docs/en/v1.15.0/pages/v2#serverbuilddirectory";
+
+export let serverBuildTargetWarning =
+  "⚠️ REMIX FUTURE CHANGE: The `serverBuildTarget` config option will be removed in v2. " +
+  "Use a combination of server module config values to achieve the same build output. " +
+  "For instructions on making this change see " +
+  "https://remix.run/docs/en/v1.15.0/pages/v2#serverbuildtarget";
+
+export const serverModuleFormatWarning =
+  "⚠️ REMIX FUTURE CHANGE: The `serverModuleFormat` config default option will be changing in v2 " +
+  "from `cjs` to `esm`. You can prepare for this change by explicitly specifying `serverModuleFormat: 'cjs'`. " +
+  "For instructions on making this change see " +
+  "https://remix.run/docs/en/v1.16.0/pages/v2#servermoduleformat";
+
+export let flatRoutesWarning =
+  "⚠️ REMIX FUTURE CHANGE: The route file convention is changing in v2. " +
+  "You can prepare for this change at your convenience with the `v2_routeConvention` future flag. " +
+  "For instructions on making this change see " +
+  "https://remix.run/docs/en/v1.15.0/pages/v2#file-system-route-convention";
 
 export const errorBoundaryWarning =
-  "⚠️ DEPRECATED: The separation of `CatchBoundary` and `ErrorBoundary` has " +
-  "been deprecated and Remix v2 will use a singular `ErrorBoundary` for " +
-  "all thrown values (`Response` and `Error`). Please migrate to the new " +
-  "behavior in Remix v1 via the `future.v2_errorBoundary` flag in your " +
-  "`remix.config.js` file. For more information, see " +
-  "https://remix.run/docs/route/error-boundary-v2";
+  "⚠️ REMIX FUTURE CHANGE: The behaviors of `CatchBoundary` and `ErrorBoundary` are changing in v2. " +
+  "You can prepare for this change at your convenience with the `v2_errorBoundary` future flag. " +
+  "For instructions on making this change see " +
+  "https://remix.run/docs/en/v1.15.0/pages/v2#catchboundary-and-errorboundary";
+
+export const formMethodWarning =
+  "⚠️ REMIX FUTURE CHANGE: APIs that provide `formMethod` will be changing in v2. " +
+  "All values will be uppercase (GET, POST, etc.) instead of lowercase (get, post, etc.) " +
+  "You can prepare for this change at your convenience with the `v2_normalizeFormMethod` future flag. " +
+  "For instructions on making this change see " +
+  "https://remix.run/docs/en/v1.15.0/pages/v2#formMethod";
+
+export const metaWarning =
+  "⚠️ REMIX FUTURE CHANGE: The route `meta` export signature is changing in v2. " +
+  "You can prepare for this change at your convenience with the `v2_meta` future flag. " +
+  "For instructions on making this change see " +
+  "https://remix.run/docs/en/v1.15.0/pages/v2#meta";