+++ packages/remix-netlify/CHANGELOG.md
@@ -1,5 +1,34 @@
 # `@remix-run/netlify`
 
+## 1.16.0
+
+### Patch Changes
+
+- feat: support async `getLoadContext` in all adapters ([#6170](https://github.com/remix-run/remix/pull/6170))
+- Updated dependencies:
+  - `@remix-run/node@1.16.0`
+
+## 1.15.0
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/node@1.15.0`
+
+## 1.14.3
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/node@1.14.3`
+
+## 1.14.2
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/node@1.14.2`
+
 ## 1.14.1
 
 ### Patch Changes